Move Device and time to new crate
